Cleaned up wallet screen by dividing into components
