+++ Screens/WalletScreen/Components/BalanceDisplay.js
@@ -1,0 +1,33 @@
+import React, { Component } from "react";
+import { Text } from "react-native-elements";
+import Styles from "../../../Styles/Styles.js";
+
+
+export default class BalanceDisplay extends Component {
+
+    constructor(props) {
+        super(props);
+    }
+
+    balanceToString() {
+        let bal = this.props.balance.toString();
+    
+        if (bal === "0") return bal;
+        else {
+          if (bal.length <= 6) {
+            return "0." + "0".repeat(6 - bal.length) + bal;
+          } else {
+            return bal.slice(0, bal.length - 6) + "." + bal.slice(bal.length - 6);
+          }
+        }
+      }
+
+    render() {
+        return (
+        <Text h2={true} h2Style={Styles.headerText}>
+            {this.balanceToString()}
+        </Text>
+        )
+    }
+
+}
